Move test file to test code
